+++ packages/cli/src/types/index.ts
@@ -1,5 +1,6 @@
 export interface ConfigContent {
   token?: string;
+  accounts?: Account[];
 }
 
 export type Config = ConfigContent | Record<string, any>;
@@ -13,4 +14,9 @@ export interface EnvVariable {
   keyValue?: string;
   projects?: string;
   environment_keyName?: string;
+}
+
+export interface Account {
+  account_id: number;
+  account_alias: string;
 }