+++ packages/cli/src/util/login/execute-verify.ts
@@ -1,19 +1,25 @@
+import { Account } from "../../types/index.js";
 import { Client } from "../client.js";
 
 export interface VerifyResult {
   token: string;
+  accounts?: Account[];
 }
 
 export default async function executeVerify(
   client: Client,
   email: string | string[],
   token: string,
-): Promise<VerifyResult> {
-  return await client.fetch<VerifyResult>(`${client.apiUrl}/onboarding/verify`, {
-    method: "POST",
-    body: {
-      email,
-      token,
-    },
-  });
+): Promise<VerifyResult | number> {
+  try {
+    return await client.fetch<VerifyResult>(`${client.apiUrl}/onboarding/verify`, {
+      method: "POST",
+      body: {
+        email,
+        token,
+      },
+    });
+  } catch (err) {
+    return 1;
+  }
 }